virtme: Introduce --shared-path argument
The expected format is:
  --shared-path tag_name=/path/on/host:/path/on/guest

This new option will start new virtiofsd instances to shared paths on
the host to the guest. The new feature makes easier to share scripts
to test from the host to test the guest when using --root option.

The paths are automounted by virtme-init and virtme-ng-init on the path
specified after the colon punctuation mark, like the how the container
engines map volumes.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>