virtme-ng: MCP: Get rid of run_kernel
Now that run_kernel_async operates in a synchronous mode for the first
60 seconds, there's no reason to have two separate commands, we can
always use run_kernel_async and the MCP server will automatically use
the synchronous mode for short commands and the asynchronous mode for
long commands.

Therefore we can get rid of run_kernel and simplify the code.

Signed-off-by: Andrea Righi <arighi@nvidia.com>